Fixed bug: More reviews now renders more review correctly.
Issue: When setting state, that new state is not yet updated so the subsequent action in the same function scope should not depend on those changes to happen.
i.e. I updated max result by adding 2 to get 4 and then tried to refer to max result, hoping to get 4 but it was return 2 (the old state), so I just used old state + 2 in the next line.